+++ shared/src/main/scala/citeobj/CiteCatalog.scala
@@ -34,8 +34,20 @@ import js.annotation.JSExport
       case 1 => Some(filtered(0))
       case _ => throw CiteObjectException("Duplicate URNs found in catalog: " + urn)
     }
-
   }
+
+  /** True if collection is ordered.
+  *
+  * @param coll Collection to test.
+  */
+  def isOrdered(urn : Cite2Urn) : Boolean = {
+    collection(urn).get.orderingProperty match {
+      case None => false
+      case seqProp : Option[Cite2Urn] => true
+    }
+  }
+
+
 
   /** Set of URNs identifying all properties in the repository.
   */