+++ releases.md
@@ -1,8 +1,9 @@
 #`citeobj`: release notes
 
+**1.2.0**: Adds functionality for working with ordered collections.
 
 **1.1.1**: Bug fixes in exposing shared code to ScalaJS environment.
 
-**1.1**: Support building a data collection from a string source in CEX format.
+**1.1.0**: Support building a data collection from a string source in CEX format.
 
-**1.0**: initial release.  Supports building a CiteCollectionRepository in various ways, including from a String in CEX 1.1 format.  Alternatively, in the JVM subproject, XML catalogs can be used with delimited text data to create a repository.
+**1.0.0**: initial release.  Supports building a CiteCollectionRepository in various ways, including from a String in CEX 1.1 format.  Alternatively, in the JVM subproject, XML catalogs can be used with delimited text data to create a repository.